Move attribute as a top-most package
