+++ pkg/attribute/checker.go
