+++ pkg/attribute/checker_test.go
