+++ pkg/sql/codegen/pai/kmeans.go
@@ -17,10 +17,10 @@ import (
 	"fmt"
 	"strings"
 
+	"sqlflow.org/sqlflow/pkg/attribute"
 	"sqlflow.org/sqlflow/pkg/database"
 	"sqlflow.org/sqlflow/pkg/ir"
 	pb "sqlflow.org/sqlflow/pkg/proto"
-	"sqlflow.org/sqlflow/pkg/sql/codegen/attribute"
 )
 
 var kmeansAttributes = attribute.Dictionary{
