+++ pkg/sql/codegen/tensorflow/codegen.go
@@ -22,9 +22,9 @@ import (
 	"strings"
 	"text/template"
 
+	"sqlflow.org/sqlflow/pkg/attribute"
 	"sqlflow.org/sqlflow/pkg/ir"
 	pb "sqlflow.org/sqlflow/pkg/proto"
-	"sqlflow.org/sqlflow/pkg/sql/codegen/attribute"
 )
 
 var commonAttributes = attribute.Dictionary{
